requirements.txt
flask>=2.0.0
pandas>=1.3.0
openpyxl>=3.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
numpy>=1.21.0
pyjwt>=2.0.0
schedule>=1.1.0
gunicorn>=21.2.0